feat: 예약이 존재하는 시간을 삭제할때의 예외 구현

+++ src/main/java/roomescape/reservationtime/service/ReservationTimeService.java
@@ -4,6 +4,8 @@ import java.time.LocalTime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import roomescape.error.ReservationException;
+import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservationtime.domain.ReservationTime;
 import roomescape.reservationtime.dto.ReservationTimeResponse;
 import roomescape.reservationtime.repository.ReservationTimeRepository;
@@ -13,6 +15,7 @@ import roomescape.reservationtime.repository.ReservationTimeRepository;
 public class ReservationTimeService {
 
     private final ReservationTimeRepository reservationTimeRepository;
+    private final ReservationRepository reservationRepository;
 
     public ReservationTimeResponse saveTime(final LocalTime time) {
         final ReservationTime reservationTime = reservationTimeRepository.save(time);
@@ -27,6 +30,11 @@ public class ReservationTimeService {
     }
 
     public void delete(final Long id) {
+        if (reservationRepository.existsByReservationTimeId(id)) {
+            throw new ReservationException("해당 시간으로 예약된 건이 존재합니다.");
+        }
+
         reservationTimeRepository.deleteById(id);
     }
 }
+
